requirements.txt
aiohttp
pyyaml
requests
